LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/text/2d.text.measure.emHeights-expected.txt

2d.text.measure.emHeights
Testing emHeights
Actual output:

FAIL Testing emHeights assert_equals: ctx.measureText('A').emHeightAscent === 37.5 (got 38[number], expected 37.5[number]) expected 37.5 but got 38
